requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --no-emit-index-url
#
asgiref==3.7.2
    # via django
certifi==2023.5.7
    # via requests
cffi==1.16.0
    # via cryptography
cfgv==3.3.1
    # via pre-commit
charset-normalizer==3.1.0
    # via requests
cryptography==41.0.5
    # via
    #   -r requirements.in
    #   secretstorage
distlib==0.3.6
    # via virtualenv
django==4.2.1
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-health-check
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
django-extensions==3.2.3
    # via -r requirements.in
django-health-check==3.18.1
    # via -r requirements.in
django-stubs==4.2.1
    # via
    #   -r requirements.in
    #   djangorestframework-stubs
django-stubs-ext==4.2.1
    # via django-stubs
djangorestframework==3.14.0
    # via -r requirements.in
djangorestframework-stubs[compatible-mypy]==3.14.0
    # via -r requirements.in
docutils==0.20.1
    # via readme-renderer
filelock==3.12.0
    # via virtualenv
identify==2.5.24
    # via pre-commit
idna==3.4
    # via requests
importlib-metadata==7.0.1
    # via
    #   keyring
    #   twine
jaraco-classes==3.3.0
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
keyring==24.3.0
    # via twine
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.2.0
    # via jaraco-classes
mypy==1.2.0
    # via
    #   -r requirements.in
    #   django-stubs
    #   djangorestframework-stubs
mypy-extensions==1.0.0
    # via mypy
nh3==0.2.15
    # via readme-renderer
nodeenv==1.8.0
    # via pre-commit
pkginfo==1.9.6
    # via twine
platformdirs==3.5.1
    # via virtualenv
pre-commit==3.3.2
    # via -r requirements.in
pycparser==2.21
    # via cffi
pygments==2.17.2
    # via
    #   readme-renderer
    #   rich
pyjwt==2.8.0
    # via -r requirements.in
pytz==2024.1
    # via djangorestframework
pyyaml==6.0.1
    # via pre-commit
readme-renderer==42.0
    # via twine
requests==2.31.0
    # via
    #   -r requirements.in
    #   djangorestframework-stubs
    #   requests-toolbelt
    #   twine
requests-toolbelt==1.0.0
    # via twine
rfc3986==2.0.0
    # via twine
rich==13.7.0
    # via twine
secretstorage==3.3.3
    # via keyring
sqlparse==0.4.4
    # via django
twine==4.0.2
    # via -r requirements.in
types-pytz==2023.3.0.0
    # via django-stubs
types-pyyaml==6.0.12.10
    # via
    #   django-stubs
    #   djangorestframework-stubs
types-requests==2.31.0.1
    # via
    #   -r requirements.in
    #   djangorestframework-stubs
types-urllib3==1.26.25.13
    # via types-requests
typing-extensions==4.6.3
    # via
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   mypy
urllib3==2.0.3
    # via
    #   requests
    #   twine
virtualenv==20.23.0
    # via pre-commit
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
